pyproject.toml
[tool.poetry]
name = "portfolio-api"
version = "0.1.0"
description = "API for portfolio data"
authors = ["tjeerddie <tjeerdverschragen@hotmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.2"
black = "^20.8b1"
flake8 = "^3.8.4"
Pillow = "^8.0.1"
python-dotenv = "^0.14.0"
django-imagekit = "^4.0.2"
djangorestframework = "^3.12.1"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

